+++ src/main/java/kitchenpos/table/validator/OrderValidatorImpl.java
@@ -1,0 +1,29 @@
+package kitchenpos.table.validator;
+
+import kitchenpos.order.dto.OrderRequest;
+import kitchenpos.order.validator.OrderValidator;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.repository.OrderTableRepository;
+import kitchenpos.table.exception.TableException;
+import kitchenpos.table.exception.TableExceptionType;
+import org.springframework.stereotype.Component;
+
+@Component
+public class OrderValidatorImpl implements OrderValidator {
+    private final OrderTableRepository orderTableRepository;
+
+    public OrderValidatorImpl(final OrderTableRepository orderTableRepository) {
+        this.orderTableRepository = orderTableRepository;
+    }
+
+    @Override
+    public void validate(final OrderRequest orderRequest) {
+        final OrderTable orderTable = findByOrderTableId(orderRequest);
+        orderTable.validateUsed();
+    }
+
+    private OrderTable findByOrderTableId(final OrderRequest orderRequest) {
+        return orderTableRepository.findById(orderRequest.getOrderTableId())
+                .orElseThrow(() -> new TableException(TableExceptionType.TABLE_NOT_FOUND));
+    }
+}
