+++ src/test/java/kitchenpos/order/validator/OrderValidatorTest.java
@@ -5,6 +5,7 @@ import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.repository.OrderTableRepository;
 import kitchenpos.table.exception.TableException;
+import kitchenpos.table.validator.OrderValidatorImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -17,7 +18,6 @@ import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 
@@ -26,7 +26,7 @@ public class OrderValidatorTest {
     @Mock
     OrderTableRepository orderTableRepository;
     @InjectMocks
-    OrderValidator orderValidator;
+    OrderValidatorImpl orderValidator;
 
     @Test
     @DisplayName("테이블 변경 가능")
